requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
paho-mqtt==1.6.1
aiofiles==23.2.1
python-multipart==0.0.9
